+++ src/Faithlife.Analyzers/IfNotNullCodeFixProvider.cs
@@ -46,7 +46,15 @@ public sealed class IfNotNullCodeFixProvider : CodeFixProvider
 		// The location of each of the arguments changes based on whether the method is invoked as an extension method.
 		var targetExpression = methodSymbol.IsStatic ?
 			ifNotNullInvocation.ArgumentList.Arguments[0].Expression :
-			((MemberAccessExpressionSyntax) ifNotNullInvocation.Expression).Expression;
+			ifNotNullInvocation.Expression switch
+			{
+				MemberAccessExpressionSyntax memberAccess => memberAccess.Expression,
+				MemberBindingExpressionSyntax memberBinding => (memberBinding.Parent.Parent as ConditionalAccessExpressionSyntax)?.Expression,
+				ElementBindingExpressionSyntax elementBinding => (elementBinding.Parent.Parent as ConditionalAccessExpressionSyntax)?.Expression,
+				_ => null,
+			};
+		if (targetExpression is null)
+			return;
 
 		var delegateExpression = methodSymbol.IsStatic ?
 			ifNotNullInvocation.ArgumentList.Arguments[1].Expression :
@@ -123,7 +131,7 @@ public sealed class IfNotNullCodeFixProvider : CodeFixProvider
 				if (!outputTypeArgument!.CanBeReferencedByName)
 					return;
 
-				defaultValueExpression = DefaultExpression(GetTypeSyntax(outputTypeArgument));
+				defaultValueExpression = LiteralExpression(SyntaxKind.DefaultLiteralExpression);
 			}
 		}
 
@@ -267,7 +275,11 @@ public sealed class IfNotNullCodeFixProvider : CodeFixProvider
 							title: "Use conditional access operator",
 							createChangedDocument: token => ReplaceValueAsync(
 								context.Document,
-								ifNotNullInvocation,
+								ifNotNullInvocation.Expression is MemberBindingExpressionSyntax or ElementBindingExpressionSyntax &&
+								ifNotNullInvocation.Parent is ConditionalAccessExpressionSyntax conditionalAccess &&
+								conditionalAccess.WhenNotNull == ifNotNullInvocation
+									? conditionalAccess
+									: ifNotNullInvocation,
 								finalExpression,
 								token),
 							c_fixName),
@@ -337,7 +349,7 @@ public sealed class IfNotNullCodeFixProvider : CodeFixProvider
 			return;
 		}
 
-		ExpressionSyntax replacementTarget;
+		SyntaxNode replacementTarget;
 		ExpressionSyntax replacementExpression;
 
 		if (defaultValueExpression is null &&
@@ -357,11 +369,15 @@ public sealed class IfNotNullCodeFixProvider : CodeFixProvider
 		}
 		else if (defaultValueExpression is object || outputTypeArgument!.CanBeReferencedByName) //// TODO: verify this null coercion is safe
 		{
-			replacementTarget = ifNotNullInvocation;
+			replacementTarget = ifNotNullInvocation.Expression is MemberBindingExpressionSyntax or ElementBindingExpressionSyntax &&
+				ifNotNullInvocation.Parent is ConditionalAccessExpressionSyntax conditionalAccess &&
+				conditionalAccess.WhenNotNull == ifNotNullInvocation
+					? conditionalAccess
+					: ifNotNullInvocation;
 			replacementExpression = ConditionalExpression(
 				conditionExpression,
 				SyntaxUtility.SimplifiableParentheses(lambdaExpressionBody),
-				defaultValueExpression ?? DefaultExpression(GetTypeSyntax(outputTypeArgument!))); // TODO: verify this null coercion is safe
+				defaultValueExpression ?? LiteralExpression(SyntaxKind.DefaultLiteralExpression)); // TODO: verify this null coercion is safe
 		}
 		else
 		{